requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
django-meta==2.4.2
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.3.0
psycopg==3.2.6
psycopg2==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.3
twilio==9.5.0
urllib3==2.3.0
yarl==1.18.3